change semantics of wiwXsismo badge
