+++ flows/playground.ts
@@ -497,7 +497,7 @@ export const playgroundFlows: Flow[] = [
     congratulationTexts: ["Congratulations you're an Active Aztec user!"],
   },
   {
-    path: "nft-degen",
+    path: "nft-legendary-traders",
     attester: hydraS1SimpleAttester.name,
     network: Network.Polygon,
     attesterType: "hydra-s1",
@@ -505,26 +505,11 @@ export const playgroundFlows: Flow[] = [
     badgesInternalCollectionsIds: [201],
     title: "WIW",
     logoUrl: "https://static.wiw.io/image/logo.png",
-    subtitle: "Prove you are a Degen NFT Trader.",
-    onboardingDescription: "You have claimed the NFT Degen ZK Badge!",
+    subtitle: "Prove you are one of the NFT Legendary Traders.",
+    onboardingDescription: "You have claimed the NFT Legendary Trader ZK Badge!",
     ctaLabel: "Explore WIW",
     ctaUrl: "<https://wiw.io>",
-    congratulationTexts: ["You have claimed the NFT Degen ZK Badge!"],
-  },
-  {
-    path: "crypto-memento",
-    attester: hydraS1SimpleAttester.name,
-    network: Network.Polygon,
-    attesterType: "hydra-s1",
-    badgesCollection: hydraS1SimpleBadges,
-    badgesInternalCollectionsIds: [202],
-    title: "WIW",
-    logoUrl: "https://static.wiw.io/image/logo.png",
-    subtitle: "Prove you have experienced crypto historical moment.",
-    onboardingDescription: "You have claimed the Crypto Memento ZK Badge — Vol.1!",
-    ctaLabel: "Explore WIW",
-    ctaUrl: "<https://wiw.io>",
-    congratulationTexts: ["You have claimed the Crypto Memento ZK Badge — Vol.1!"],
+    congratulationTexts: ["You have claimed the NFT Legendary Trader ZK Badge!"],
   },
   {
     path: "timeswap-lens-followers",
