+++ group-generators/generators/index.ts
@@ -94,8 +94,7 @@ import tutoEnsContributors from "./tuto-ens-contributors"
 import twitterEthereumInfluencers from "./twitter-ethereum-influencers"
 import wagameLensPostInteraction from "./wagame-lens-post-interaction";
 import walletGroup from "./wallet-group";
-import wiwCryptoMemento from "./wiw-crypto-memento";
-import wiwNftDegenTraders from "./wiw-nft-degen-traders";
+import wiwNftLegendaryTraders from "./wiw-nft-legendary-traders";
 import zkHackIiiSismoWorkshop from "./zk-hack-iii-sismo-workshop";
 
 export const groupGenerators: GroupGeneratorsLibrary = {
@@ -193,7 +192,6 @@ export const groupGenerators: GroupGeneratorsLibrary = {
   "wagame-lens-post-interaction": wagameLensPostInteraction,
   "rhinofi-power-users": rhinofiPowerUsers, 
   "wallet-group": walletGroup,
-  "wiw-crypto-memento": wiwCryptoMemento,
-  "wiw-nft-degen-traders": wiwNftDegenTraders,
+  "wiw-nft-legendary-traders": wiwNftLegendaryTraders,
   "zk-hack-iii-sismo-workshop": zkHackIiiSismoWorkshop,
 };
